% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockshap_accessors.R
\name{clockshap_deviation}
\alias{clockshap_deviation}
\title{Extract ClockSHAP deviation values}
\usage{
clockshap_deviation(x)
}
\arguments{
\item{x}{A \code{clockshap} object.}
}
\value{
A numeric vector of deviations.
}
\description{
Extract ClockSHAP deviation values
}
