% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockshap_accessors.R
\name{clockshap_expected}
\alias{clockshap_expected}
\title{Extract ClockSHAP expected ages}
\usage{
clockshap_expected(x)
}
\arguments{
\item{x}{A \code{clockshap} object.}
}
\value{
A numeric vector of expected ages.
}
\description{
Extract ClockSHAP expected ages
}
