% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockshap_accessors.R
\name{clockshap_phi}
\alias{clockshap_phi}
\title{Extract ClockSHAP feature contributions}
\usage{
clockshap_phi(x)
}
\arguments{
\item{x}{A \code{clockshap} object.}
}
\value{
A numeric matrix of feature contributions
(samples x features).
}
\description{
Extract ClockSHAP feature contributions
}
