% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockshap_accessors.R
\name{clockshap_predicted}
\alias{clockshap_predicted}
\title{Extract ClockSHAP predicted ages}
\usage{
clockshap_predicted(x)
}
\arguments{
\item{x}{A \code{clockshap} object.}
}
\value{
A numeric vector of predicted ages.
}
\description{
Extract ClockSHAP predicted ages
}
